refactor: 경매 상품 상세페이지 파일 분리
