+++ apps/web/features/auction/detail/ui/AuctionDetail.tsx
